Allow to specify notification's hints
Desktop Notifications Specification [1] specifies that hints can be
used to provide extra data to a notification server.

A new command line option --hint allows to add a hint to the
notification to display.
This option can be used multiple times, one for each hint to add.
--hint option format is name:value.

The new 'hints' command allow to specify hints in 'listen' mode.
Same format that in the command line option is used.
Several hints can be passed by separating them by '\n'.

Hints of value type '(iiibiiay)' are not supported.
This value type is used to pass a raw data image as a hint value.

This new change is useful for implementing the NotificationSource [2]
GNOME Goal.
A application using zenity and having a desktop file, can now specify
that it is a notification emitter and it's notifications can be
filtered in the new Notifications GNOME control panel pane.

[1] http://people.gnome.org/~mccann/docs/notification-spec/notification-spec-latest.html#hints
[2] https://live.gnome.org/GnomeGoals/NotificationSource

https://bugzilla.gnome.org/show_bug.cgi?id=693751