l10n updates and update manpage and yelp.
l10n po-files regenerated. This has resulted in substantial fuzzy
strings that will require translator attention.

I have updated the en_CA translation myself.

I have also updated the manpage and 'C' locale of the Yelp help to
reflect the change in options, particularly to the removal of the
'window-icon' option and moving it to the 'icon' option which is
specific to the libnotify support now.