msg: Fix maximum width to 60 characters
Commit 3b64d05e8a1aae1581d2bec7288d80ac4699e1b1 mentioned that labels in
message dialogs were set to a maximum width of 60 characters, with
inspiration from GtkMessageDialog, but in fact a typo meant that they
were set to both a width and maximum width of just 10 characters. This
lead to text layout problems, which can be fixed by using the intended
maximum of 60 characters.

https://bugzilla.redhat.com/show_bug.cgi?id=2053530