Set dialog 'heading', not 'title'
In the GTK 4 port, we made all Zenity dialogs use the libadwaita
`AdwMessageDialog` class. This dialog does not have a title bar
and does not show its title. The equivalent to title for this
class is 'heading' - everywhere we were previously setting the
title, we should set the 'heading' instead.

Signed-off-by: Adam Williamson <awilliam@redhat.com>