tree-column-view: don't steal children
When mousing over a dialog that uses this kind of view (e.g.
one created with the test command in
https://gitlab.gnome.org/GNOME/zenity/-/issues/54#note_1726876 ),
I see a lot of these errors on the console:
gtk_list_item_set_child: assertion 'child == NULL || gtk_widget_get_parent (child) == NULL' failed
I think we can hit this callback more than once, and doing so
means we try to set the list item's child again; Gtk doesn't like
that, because we're effectively trying to set the parent of an
item that already has one (it doesn't check if we're just trying
to set the *same* parent, it just complains any time you try to
set something as the child that already has a parent).

Signed-off-by: Adam Williamson <awilliam@redhat.com>