Port to webkitgtk 6.0
webkitgtk 5.0 is gone from Fedora 38 and Fedora Rawhide, so our
CI 'build_fedora' task is always failing. This ports the code to
webkitgtk 6.0 and updates the build and CI configs. I based the
changes on Epiphany code. I tested it with:

zenity --text-info --html --url=https://www.happyassassin.net

and it seems to work OK.

Signed-off-by: Adam Williamson <awilliam@redhat.com>