data: Manpage generation changes
Don't automatically generate the manpage from help2man for now. This
creates problems since a display manager is required to run `zenity
--help-all` and `zenity --version` at this time. This should be fixed in
the long term, but for now we're just going to avoid the issue
altogether.

Manual intervention will be required to update the manpage for now, but
hopefully we can rectify this in the future. It is fine for now as the
manpage should not need to be regenerated too frequently.

Also, s/GTK+/GTK in the .roff boilerplate.