Fix --extra-button and delete events for legacy (non-libadwaita) dialogs
ie, they no longer print junk and/or crash upon delete events.

This only worked in 3.x due to a lucky fluke of comparisons of mixed
signedness happening to work out and warnings being suppressed.

Also, add deprecation warning for --extra-buttons in the context of
--file-selection. We don't want scripts to fail to run since it worked
in older versions of 3.x, but a warning message now more consistent with
other deprecation warnings will be displayed.