tree: Set search column to 1 for check/radiolists
We expand 91a6c185 to apply to check/radiolists as the first column
(ie, true/false) gets selected as the search column by default unless it
is explicitly set.

nb: Although acc. to the GtkTreeView docs, the default property for
'search-column' is -1 (ie, disable search) it seems that gtktreeview.c
will auotmatically select the first available column as being searchable
(ie, 0) so it is not necessary to explicitly set this property to 0 in
other non-image/radio/checklist cases.

Fixes #49