option: Don't pass around argv vector from main which is likely on stack
Although the C11 standard guarantees that the strings themselves from
the argv vector will be guaranteed to remain for the life of the
program, that is not so with the vector itself passed to main.

Use a duplicated GStrv to avoid this.

This is likely an oversight from the port to GApplication, where it is
likely undefined behaviour to use the argv vector in a callback which
may not be in the same stack frame as main by the time we get there.

Likely fixes RH bugs:
https://bugzilla.redhat.com/show_bug.cgi?id=2216049
https://bugzilla.redhat.com/show_bug.cgi?id=2214000

Also, avoid useless GSList usage in entry.c when parsing the 'data'
option parameters. Could also likely be avoided in tree.c, but there is
more work to be done there to eliminate it, so there is no urgency.