Add rudimentary test suite.
Perl5 is required. As Zenity is a GUI/command-line application, and many
of the tests require user input/intervention, these kinds of tests
cannot be integrated into a CI environment and must be run manually.

This is just an initial set of unit tests for now, and as we think of
more, I'm sure many more can be added.

TODO: At least integrate into meson by allowing 'meson test' to run the
test suite.