forms: Port GtkComboBox to GtkDropDown
Still need to port GtkTreeView. But the test-suite was revealing a
GObject error upon exit with GtkComboBox, so I decided to just port to
GtkDropDown and get it done with. The GObject error goes away and we
also got rid of a lot of gross old code, so it's a win-win.