feat(progress): support markup on label update STDIN
This commit adds support for markup when updating the progress bar
label with values read from STDIN.

Initially, commit 70ea28f3 introduced reading from STDIN. At that
time, markup support was not available, as neither the initial setting
nor updates via STDIN utilized it. Later, commit 5547de93 introduced
support for markup, but only for the initial label. There are no
explanations in the commit message for this limitation, so it seems
unintentional.

Supporting markup on every update of the label enhances the usability
of the progress bar. The label can be used for structured messages,
which may improve the visibility of long texts for the user.

fix issue #93