Disable AdwMessageDialog deprecation warnings for libadwaita >= 1.6
During the zenity4 development cycle, the app was already rewritten
twice to (a) port to GtkWindow after GtkDialog was deprecated; (b) port
to AdwMessageDialog as the shiny new dialog API in libadwaita 1.2 when
it was released. And now for libadwaita 1.6 AdwMessageDialog is
deprecated. We're not doing a 3rd rewrite at this time.

Add function to util.c to make this process easier.